22 sipm hv control (#72)
* Added PDUMapping and Change SPIInterfaceMultiplexer

* Refactored relationship between SPIInterface and SPIInterfaceMultiplexer

* Added GetPDUInfo and ControlPDU

* added pdu test code

* change definitions of argument type in SPI functions

* debug message inserted

* Bug fix in MPSSEController.cc

* bug fix in MPSSE

* Change test code for PDU to use Ft232H

* bug fix in pdu control

* bug fix in pdu control

* added dkms tools

* Put back the BME680 library

* Refactored VersaAPI calling method

* Refactored to divide versalogic API functions to multiple classes

* Added Pressure Telemetry

* add pdu_test code

* delete BME680 api

* modified the serialize and initializeDBTable. And defined setter and getter in HubHKTelemetry

* add TofBiasController

* Add some member function to TofBiasController

* bug fix in MPSSE

* add pdu read example

* Add utility function to make option

* improve test codes

* Compile-time checking of range in HubHKTelemetry for VHKtelemetryMapping

* Introduce DBSerializable

* Modified interpret() in HubHKTelemetry

* update HubHKTelemetry::update()

* Added printIterative function in HubHKTelemetry

* Converting telemetry values and updated serialize function in HubHKTelemetry

* Bug fix in compiling

* Change inclinometer mapping

* tweaks

* Bug fix in ConvertedHubHKTelemetry and added conversion about hub computer

* Bug  fix in inlinometer readout

* added some conversion formula

* Add ToFBiasChannel and commanding for ToF bias.

* fix compile error on hub computer

* update PDU mapping

* increase PDU CS mapping for telemetry

* Introduce VDBDataStore class for DataBase treatment

* update interface file

* bug fix in telemetry interpret classes

* modified monitor.rb to push converted data to mysql

* tweaks

* add getALLAdcs in ADC io class

* Change implementation on PDUCSmapping and added PDUChannelMap

* Implement Control of DAC

* Bug fix in SPIInterfaceMultiplexer

* Bug fix and preparation for PDU commanding

* fix compile error

* Adjust the include files

* tweaks

* tweaks

* Tweaks

* add member function of setSpiMode

* bug fix in BaycatSPIIO

* bug fix in FPGAGPIO

* Bug fix in PDU multiplexer channel mapping

* Bug fix in PDUCSMapping

* Added PDU Code Map and implemented mapping and commands

* bug fix in SencTelemetry and ReceiveCommand and update pdu_test.cc

* change implementations of SPIManager::controGPIO

* added setConfigOptions to SPIInterfaceMultiplexer and fixed pdu test code

* changed .clang-format

* add readme and tweaks

* Test done

* Implemented the Command sender

* Update Command Definition

* Bug fix in SPI Manager

* preparation for MCP2210

* bug fix

* tweaks

* Squashed commit of the following:

commit b59c5d89be3c2f75c87a39e901414a994a8c5838
Author: ShotaArai <93177954+STA205233@users.noreply.github.com>
Date:   Sat Aug 8 00:03:13 2026 +0900

    56 change in labjack readout (#57)

    * change labjack setting for differential readout

    * Tweaks

commit fc53d39063f9da22f4ab4dbddb7a1dd7ab7b996c
Author: ShotaArai <93177954+STA205233@users.noreply.github.com>
Date:   Thu Aug 6 17:14:53 2026 +0900

    45 multi threaded mqtt communication (#54)

    * use of loop_start and loop_end for MT MQTT communication

    * Delete threaded set parameters

    * bug fix in stopping MQTT thread

    * Use of connect_async and introduce loop() to make sure connection is successful.

    * use of static member variable to manage API initialization

    * Revert "use of static member variable to manage API initialization"

    This reverts commit e5da2bab88b6c3a08216b62951bbcbcf97368879.

    * tweaks

    * bug fix in MosquittoIO

    * update PDU command in CommandBuilder

    * fix to match new version ComminucationCodes.hh

    * surpress verbose level and change default exec to hubcomputer.rb

    * tweaks

    * tweaks

commit 7fc37ddabc970ac54b12ec9fd2d7759c6344b0b7
Author: Shota Arai <shota.arai@phys.s.u-tokyo.ac.jp>
Date:   Tue Aug 4 10:16:01 2026 -0400

    bug fix in sendtelemetry

commit 911c84d650f734186f5b3afc97b97b66e58f126f
Author: ShotaArai <93177954+STA205233@users.noreply.github.com>
Date:   Sat Jul 25 21:02:54 2026 +0900

    feature/docs (#52)

    * Update installation documentation for pGramsFC, including prerequisite software and installation procedures

    * Merge branch 'master' into feature/docs

    * Merge remote-tracking branch 'origin/master' into feature/docs

    * Added to MakeDocument.rb / added argument loader to CommandSender.md

    * Merge branch 'master' into feature/docs

    * update documentation

    * Change description of classes

    * Merge branch 'master' into feature/docs

commit 0097f3bec7305d71ee8d8d0d3c9470b4c4a0c805
Author: ShotaArai <93177954+STA205233@users.noreply.github.com>
Date:   Sat Jul 25 17:46:00 2026 +0900

    Feature/cmake (#51)

    * add clangd filea and refresh CMakeLists.txt

    * Merge branch 'master' into feature/cmake

commit b2f11ef91cae0d9ae149095ef8b8273767a0eb1e
Author: ShotaArai <93177954+STA205233@users.noreply.github.com>
Date:   Sat Jul 25 16:23:50 2026 +0900

    48 apply change of communciationcodeshh (#50)

    * update PDU command in CommandBuilder

    * fix to match new version ComminucationCodes.hh

* fix compile error

* CMakeLists

* Complete preliminary implementation for MCP2210IO.

* fix CMake  and tweaks

* Bug fix in SendTelemetry and GetPDUInfor

* put back CMakeLIsts and debug BaycatSPIIO

* preparation for MCP2210 test

* Added enable examples

* fix bug in enable test code

* Fix bug on SPI Interface

* tweaks

* Add diagnosis message when error in GPIO Control

* Bug fix in applying commands

* delete MT feature

* 58 pdu new chip (#66)

* fix mungling error

* Checked MCP2210 function

* Added MCP2210IO to SPIManager

* Refactored SerialCommunication

* fix compile error

* implemented ToF Bias Control command and telemetry
And change definition of current telemetry type

* Added max length to EncodedSerialCommunication & delete some modules

* tweaks

* Bug fix in GetPDUInfo and ADC on PDU readout

* Add TOF BIas on command

* update tof bias controller

* tweaks

* introduce monitor output

* Revert "introduce monitor output"

This reverts commit f8759a935988bb9b7dfbdbbf34af21485bcb5ed9.

* added leak test

* Delete unnecessary modules and introduce SetHKEvs and VHousekeepingModule for time management

* Bug fix in SetHKEvs and tweaks

* modified hubcomputer.rb and tweaks

* Bug fix in ToFBiasController

* ToFBiasTelemetry

* fix bug

* Delete printData and bug fix

* fix bug

* bug fix in ControlTOFBIas

* bug fix

* bug fix

* lazy readout of full output

* tweaks

* tweaks

* interpret tof bias telemetry

* fixed in conversion of MHADC

* tweaks

* implement TB_SET_V_DEF

* Bug fix in Controlling the tof bias control

* Add command to Command Builder

* added Tof bias command to GUI

* Treatment of the output of Voffset

* Implemented to TofBias displaying (not testing)

* debugging but not compiled

* Implemented displaying TB telemetry

* Bug fix

* tweaks

* Add interpretation of TOF bias control telemetry

* Bug fix in ControlToFBIas and no heartbeat sending feature

* Fix MySQL

+++ command_sender/source/send_command_raw.cc
@@ -15,9 +15,9 @@ int main(int argc, char *argv[]) {
   }
 
   uint16_t code = std::stoi(argv[2]);
-  std::vector<int32_t> arg_array;
+  std::vector<uint32_t> arg_array;
   for (int i = 3; i < argc; i++) {
-    arg_array.push_back(std::stoi(argv[i]));
+    arg_array.push_back(std::stoul(argv[i]));
   }
   
   const int result_init = mosqpp::lib_init();
