+++ onboard/source/modules/include/InterpretTelemetry.hh
@@ -50,7 +50,7 @@ public:
   anlnext::ANLStatus mod_finalize() override;
 
   std::shared_ptr<const ErrorManager> getErrorManager() const { return singleton_self()->errorManager_; }
-  int CurrentTelemetryType() const { return singleton_self()->currentTelemetryType_; }
+  uint16_t CurrentTelemetryType() const { return singleton_self()->currentTelemetryType_; }
   std::string_view TelemetryType() {
     return singleton_self()->telemetryTypeStr_;
   }
@@ -87,7 +87,7 @@ private:
   PushToMongoDB *pusher_ = nullptr;
 #endif // USE_HSQUICKLOOK
   std::shared_ptr<CommunicationSaver<std::string>> telemetrySaver_ = nullptr;
-  int currentTelemetryType_ = 0; // 1: normal, 2: tof callback
+  uint16_t currentTelemetryType_ = 0; // following the CommunicationCodes.hh
   std::map<int, std::pair<int, int>> fileIDmp_;
   bool saveTelemetry_ = true;
   std::string binaryFilenameBase_ = "";
